+++ 2DPhysicsEngine/include/PhysicShape.h
@@ -15,7 +15,7 @@ public:
 
 	[[nodiscard]] virtual ShapeType GetShapeType() const = 0;
 
-	[[nodiscard]] double GetBoundingCircleRad() const { return boundingCircleRadius_; }
+	[[nodiscard]] float GetBoundingCircleRad() const { return boundingCircleRadius_; }
 
 	/// <summary>
 	/// Computes the furthest point in a given direction
@@ -25,5 +25,5 @@ public:
 	[[nodiscard]] virtual Vector2D GetFurthestPoint(Vector2D dir) const = 0;
 
 protected:
-	double boundingCircleRadius_ = 0.0;
+	float boundingCircleRadius_ = 0.0;
 };
