+++ 2DPhysicsEngine/include/Rigibody.h
@@ -7,11 +7,11 @@ class Rigibody
 public:
 	explicit Rigibody(std::unique_ptr<PhysicShape>);
 	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos);
-	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, double mass);
+	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, float mass);
 	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, Vector2D vel);
-	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, Vector2D vel, double mass);
-	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, Vector2D vel, double mass, bool isStatic);
-	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, Vector2D vel, double mass, bool isStatic, bool useGravity);
+	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, Vector2D vel, float mass);
+	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, Vector2D vel, float mass, bool isStatic);
+	Rigibody(std::unique_ptr<PhysicShape>, Vector2D pos, Vector2D vel, float mass, bool isStatic, bool useGravity);
 
 	Rigibody(const Rigibody&) = delete;
 
@@ -20,12 +20,12 @@ public:
 	//static bool CheckCollide(Rigibody a, Rigibody b);
 	//static bool CheckBoundingBoxCollide(Rigibody a, Rigibody b);
 
-	void Update(double time);
+	void Update(float time);
 
 	void AddForceImpulse(Vector2D force);
-	void AddForceContinuous(Vector2D force, double time);
+	void AddForceContinuous(Vector2D force, float time);
 	void AddAcceleration(Vector2D acceleration);
-	void AddAccelerationContinuous(Vector2D acceleration, double time);
+	void AddAccelerationContinuous(Vector2D acceleration, float time);
 
 	[[nodiscard]] Vector2D LocalToWorldPos(Vector2D) const;
 	[[nodiscard]] Vector2D WorldToLocalPos(Vector2D) const;
@@ -33,7 +33,7 @@ public:
 	[[nodiscard]] long GetId() const { return id_; }
 	[[nodiscard]] Vector2D GetPos() const { return position_; }
 	[[nodiscard]] Vector2D GetVelocity() const { return velocity_; }
-	[[nodiscard]] double GetMass() const;
+	[[nodiscard]] float GetMass() const;
 	[[nodiscard]] PhysicShape* GetCollider() const { return colliderPtr_.get(); }
 	[[nodiscard]] bool IsStatic() const { return isStatic_; }
 	[[nodiscard]] bool UseGravity() const { return useGravity_; }
@@ -41,7 +41,7 @@ public:
 
 	void SetPos(const Vector2D newPos) { position_ = newPos; }
 	void SetVelocity(const Vector2D newVelocity) { velocity_ = newVelocity; }
-	void SetMass(const double newMass) { mass_ = newMass; }
+	void SetMass(const float newMass) { mass_ = newMass; }
 	void SetCollider(std::unique_ptr<PhysicShape> newCol) { colliderPtr_.reset(); colliderPtr_ = std::move(newCol); }
 	void SetIsColliding(const bool colliding) { isColliding_ = colliding; }
 
@@ -51,10 +51,10 @@ private:
 
 	Vector2D position_ = Vector2D(0.0, 0.0);
 	//Angle in radians
-	double rotation = 0.0;
+	float rotation = 0.0;
 	Vector2D velocity_ = Vector2D(0.0, 0.0);
 
-	double mass_ = 1.0;
+	float mass_ = 1.0;
 	bool useGravity_ = true;
 	bool isStatic_ = false;
